wire figma design in page context
