+++ assets/src/hooks/keep_scroll_at_bottom.ts
@@ -1,0 +1,10 @@
+export const KeepScrollAtBottom = {
+  updated() {
+    const messageContainer = document.querySelector('[role="message container"]');
+    if (messageContainer) {
+
+      messageContainer.scrollTop = messageContainer.scrollHeight;
+
+    }
+  },
+};
