+++ lib/oli_web/components/layouts/student_delivery.html.heex
@@ -24,5 +24,4 @@
       <%= @inner_content %>
     </div>
   </main>
-  <Components.Delivery.Layouts.footer current_user={@current_user} />
 </div>
